% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_funs.R
\name{AE_sim}
\alias{AE_sim}
\title{RAR simulator}
\usage{
AE_sim(trial, nt, theta_ae, response.type)
}
\arguments{
\item{trial}{object of class \code{trial}}

\item{nt}{number of treatments - should be teh same as in \code{trial}}

\item{theta_ae}{effect on secondary response}

\item{response.type}{response type, either 'rate', 'count' or 'absolute'}
}
\value{
updated posterior samples
}
\description{
This function simulates a RAR trial
}
