% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_funs.R
\name{alpha_compute}
\alias{alpha_compute}
\title{RAR simulator simulator}
\usage{
alpha_compute(
  nt,
  nb = 1,
  maxN = 500,
  N = 1000,
  upper = 0.95,
  lower = 0.05,
  burn = 10 * nt,
  response.type,
  conjugate_prior = T,
  padhere = padhere,
  adapt = T,
  con = T,
  M = 100
)
}
\arguments{
\item{nt}{Number of treatment arms}

\item{nb}{batch size for each update}

\item{maxN}{maximum number of patients that the trial can run for; the trial is stopped when \code{maxN} is achieved}

\item{N}{Number of particles}

\item{upper}{upper threshold for probability of superiority; the trial is stopped when one of the trial arms reaches
this value}

\item{lower}{lower threshold for probability of superiority; an arm is stopped when its corresponding probability of
superiority falls below this value}

\item{burn}{the size of pre-adaptation sample; the default is 10 times number of treatments}

\item{response.type}{response type, either 'rate', 'count' or 'absolute'}

\item{conjugate_prior}{logical argument, default is True}

\item{padhere}{vector of adherence rates for each treatment}

\item{adapt}{logical; if TRUE randomization probabilities are adapted}

\item{con}{logical; if TRUE the first arm is treated as control}
}
\value{
estimated power and type I error rate
}
\description{
This function simulates multiple RAR trials to calculate power and type I error rate
}
