% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_funs.R
\name{alpha_compute_RCT}
\alias{alpha_compute_RCT}
\title{RCT type I error rate estimation}
\usage{
alpha_compute_RCT(
  nt,
  maxN = 500,
  N = 1000,
  upper = 0.95,
  response.type,
  conjugate_prior = T,
  M = 500
)
}
\arguments{
\item{nt}{Number of treatment arms}

\item{maxN}{maximum number of patients that the trial can run for; the trial is stopped when \code{maxN} is achieved}

\item{N}{Number of particles}

\item{upper}{upper threshold for probability of superiority; the trial is stopped when one of the trial arms reaches
this value}

\item{response.type}{response type, either 'rate', 'count' or 'absolute'}

\item{conjugate_prior}{logical argument, default is True}
}
\value{
estimated type I error rate
}
\description{
This function simulates multiple RCT trials to estimate type I error rate
}
