% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_funs.R
\name{control_sup_check}
\alias{control_sup_check}
\title{Probabilities of superiority to control}
\usage{
control_sup_check(theta, con = 1)
}
\arguments{
\item{theta}{vector of (sample) effect sizes}
}
\value{
vector of probabilities of superiority to control - first component is always zero
}
\description{
This function estimates the probability of seperiority to control (first arm) for each treatment arm from a posterior sample of effect sizes
}
