% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_funs.R
\name{Gibbs}
\alias{Gibbs}
\title{MCMC step}
\usage{
Gibbs(theta, y, x, qt, response.type, ng, ...)
}
\arguments{
\item{theta}{vector of effect sizes, length equals number of arms}

\item{y}{vector of responses, length equals number of patients}

\item{x}{design matrix}

\item{qt}{proposal standard deviation}

\item{response.type}{response type, either 'rate' or 'count'}

\item{ng}{number of MCMC steps for each particle. It can be as small as 1, but need to be larger if batch updates are made}

\item{...}{covariate related arguments, \code{beta}: vector of coefficients, length equals number of covariates,
\code{z}: covariate matrix, \code{lp_beta}: a function returning log-prior for coefficients; the default is lp_default which is a diffuse normal prior,
but can be replaced with a user-specified functin, \code{qb}: proposal variance for \code{beta}}
}
\value{
vector of updated parameters
}
\description{
This function perfomrs the nested MCMC steps required in SMC
}
