% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_funs.R
\name{llik_poisson}
\alias{llik_poisson}
\title{Poisson likelihood}
\usage{
llik_poisson(theta, y, x, ...)
}
\arguments{
\item{theta}{vector of effect sizes, length equals number of arms}

\item{y}{vector of count responses, length equals number of patients}

\item{x}{design matrix}

\item{...}{covariate related arguments, \code{beta}: vector of coefficients, length equals number of covariates,
\code{z}: covariate matrix}
}
\value{
log-likelihood value
}
\description{
Returns the log-likelihood for a vector of parameter values, count responses and design matrix
}
