% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_funs.R
\name{lp_default}
\alias{lp_default}
\title{Prior}
\usage{
lp_default(par, ...)
}
\arguments{
\item{par}{vector of parameter values}

\item{...}{the normal prior hyper-parameters (optional), can be specified as vectors of the same size as \code{par} or
scalars. The default is \code{pmean = 0} and \code{psd = 10}.}
}
\value{
log-prior value
}
\description{
Returns log-prior for a vector of parameter values - this is the default normal prior function in case the user does not want to
bother to decide on a prior distribution
}
