% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_funs.R
\name{post_gamma}
\alias{post_gamma}
\title{Closed form Gamma Posterior}
\usage{
post_gamma(yb, xb, prior)
}
\arguments{
\item{yb}{vector of new responses, length equals number of patients in the most recent batch}

\item{xb}{design matrix for the most recent batch}

\item{prior}{matrix of Gamma parameters up to the most recent batch, number of rows is the
number of arms.}
}
\value{
matrix of parameters of the Gamma posterior
}
\description{
Returns updated Beta posterior parameters for the rate of success
}
