% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_funs.R
\name{post_Gauss}
\alias{post_Gauss}
\title{Closed form Gaussian Posterior}
\usage{
post_Gauss(yb, xb, prior, ysd)
}
\arguments{
\item{yb}{vector of new responses, length equals number of patients in the most recent batch}

\item{xb}{design matrix for the most recent batch}

\item{prior}{matrix of Gaussian parameters up to the most recent batch, number of rows is the
number of arms, first column contains the means and second column the sd's.}

\item{ysd}{observation variance.}
}
\value{
matrix of means and sd's for the Gaussian posterior
}
\description{
Returns updated Gaussian posterior parameters for the effect sizes
}
