% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_funs.R
\name{power_compute_RCT}
\alias{power_compute_RCT}
\title{RCT power estimation}
\usage{
power_compute_RCT(
  nt,
  theta0,
  maxN = 500,
  N = 1000,
  upper = 0.95,
  response.type,
  conjugate_prior = T,
  padhere = c(0.95, 0.95, 0.95),
  M = 500
)
}
\arguments{
\item{nt}{Number of treatment arms}

\item{theta0}{vector of effect sizes}

\item{maxN}{maximum number of patients that the trial can run for; the trial is stopped when \code{maxN} is achieved}

\item{N}{Number of particles}

\item{upper}{upper threshold for probability of superiority; the trial is stopped when one of the trial arms reaches
this value}

\item{response.type}{response type, either 'rate', 'count' or 'absolute'}

\item{conjugate_prior}{logical argument, default is True}

\item{padhere}{vector of adherence rates for each treatment}
}
\value{
estimated power
}
\description{
This function simulates multiple RCT trials to estimate power
}
