% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_funs.R
\name{sampsize}
\alias{sampsize}
\title{Sample size}
\usage{
sampsize(pars, power = 0.8, alpha = 0.05, dropout = 0.2, type, sigmas)
}
\arguments{
\item{pars}{vector of effect sizes, length equals number of arms}

\item{power}{required statistical power}

\item{alpha}{required type I error rate}

\item{type}{effect type "absolute" or "rate"}
}
\value{
sample size per arm \code{n} and total sample size \code{N}
}
\description{
This function calculates sample size for a conventional multi-arm RCT
}
