% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3arm_sim.R
\name{scn}
\alias{scn}
\title{Main simulation function}
\usage{
scn(
  i,
  cRates = c(0.02, 0.035, 0.05),
  RRR = c(0.4, 0.3),
  nb = 1000,
  maxN = 18000,
  N = 1000,
  upper1 = 0.99,
  upper2 = 0.995,
  lower = 0.01,
  futb = 0.2,
  padhere = rep(1, nt),
  adapt = F,
  control_sup = T,
  dos = 2,
  dof = 2
)
}
\arguments{
\item{i}{simulation iteration index}

\item{cRates}{vector of control risks}

\item{RRR}{vector of relative risk reductions - same size as the number of interventions}

\item{nb}{batch size; interim analysis is done at every \code{nb}patients}

\item{maxN}{maximum sample size; trial stops when \code{maxN} is reached}

\item{N}{size of posterior sample used}

\item{upper1}{superiority threshold for LP}

\item{upper2}{superiority threshold for BI}

\item{lower}{futility probability threshold}

\item{futb}{futility effect threshold (MID)}

\item{padhere}{vector of probabilities of adherence for each arm}

\item{adapt}{logical, indicating if allocation probabilities are adapted}

\item{control_sup}{logical, indicating if pairwise comparison with control is employed. If FALSE all arms are compared simultaneously}

\item{dos}{the outcome wrt which superiority decision is made taking values 1(s), 2(p1), or 3(p2)}

\item{dof}{the outcome wrt which futility decision is made taking values 1(s), 2(p1), or 3(p2)}
}
\value{
a list of simulation outputs including \code{Nt} number of arms,
        \code{early} probability of stopping early, \code{events} propostion of sepsis cases,
        \code{supstop_lp} probability of stopping due to superiority of LP,
        \code{supstop_bi} probability of stopping due to superiority of BI,
        \code{futstop_lp} probability of stopping due to futility of LP,
        \code{futstop_bi} probability of stopping due to futility of BI,
        \code{reachmax} probability of reaching max sample size,
        \code{RRR_hat} relative risk reduction estimates (posterior means),
        \code{pvalue_lp} p-values for LP, \code{pbvalue_bi} p-values for BI
}
\description{
Returns simulation results for one iteration
}
