% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3arm_sim.R
\name{sim_exe}
\alias{sim_exe}
\title{Wrapper for \code{sim_scn}}
\usage{
sim_exe(cases, M, filename)
}
\arguments{
\item{cases}{a data frame containing simulation scenarios}

\item{M}{number of simulation scenarios}

\item{filename}{file name under which the results are saved}
}
\value{
saves results ot home directory
}
\description{
applies \code{sim_scn} to multiple simulation scenarios and saves results in home directory
}
