% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3arm_sim.R
\name{sim_scn}
\alias{sim_scn}
\title{Wrapper for scn}
\usage{
sim_scn(s, p1, p2, RRR1, RRR2, upper2, M = 100)
}
\arguments{
\item{s}{ctrl risk for s}

\item{p1}{ctrl risk for p1}

\item{p2}{ctrl risk for p2}

\item{RRR1}{relative risk reduction for LP}

\item{RRR2}{relative risk reduction for BI}

\item{upper2}{superiority threshold for BI}

\item{M}{number of simulation iterations}
}
\value{
a data frame with the same output as \code{scn} for \code{M} iterations
}
\description{
returns simulation results for \code{M} iterations
}
