% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_funs.R
\name{sim_wrapper}
\alias{sim_wrapper}
\title{RAR simulator wrapper}
\usage{
sim_wrapper(
  i,
  nt,
  theta0,
  nb = 1,
  maxN = 500,
  N = 1000,
  upper = 0.95,
  lower = 0.05,
  burn = 10 * nt,
  response.type,
  conjugate_prior = T,
  padhere = rep(1, nt),
  adapt = F,
  con = T,
  prand = rep(1/nt, nt)
)
}
\arguments{
\item{i}{counter}

\item{nt}{Number of treatment arms}

\item{theta0}{vector of effect sizes}

\item{nb}{batch size for each update}

\item{maxN}{maximum number of patients that the trial can run for; the trial is stopped when \code{maxN} is achieved}

\item{N}{Number of particles}

\item{upper}{upper threshold for probability of superiority; the trial is stopped when one of the trial arms reaches
this value}

\item{lower}{lower threshold for probability of superiority; an arm is stopped when its corresponding probability of
superiority falls below this value}

\item{burn}{the size of pre-adaptation sample; the default is 10 times number of treatments}

\item{response.type}{response type, either 'rate', 'count' or 'absolute'}

\item{conjugate_prior}{logical argument, default is True}

\item{padhere}{vector of adherence rates for each treatment}

\item{adapt}{logical; if TRUE randomization probabilities are adapted}

\item{con}{logical; if TRUE the first arm is treated as control}

\item{prand}{vector of initial randomization probabilities}
}
\value{
power and alpha indicators for one trial
}
\description{
RAR simulator wrapper
}
