% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_funs.R
\name{SMC_step}
\alias{SMC_step}
\title{SMC update}
\usage{
SMC_step(theta, y, x, N, response.type = "rate", ...)
}
\arguments{
\item{theta}{matrix of posterior samples for effect sizes, columns: arms, rows: samples}

\item{y}{vector of recent responses}

\item{x}{design matrix for the recent observations}

\item{N}{number of particles}

\item{response.type}{response type, either 'rate' or 'count'}

\item{...}{covariate related arguments, \code{beta}: vector of coefficients, length equals number of covariates,
\code{z}: covariate matrix}
}
\value{
updated posterior samples
}
\description{
This function updates a posterior sample of effect sizes according to newly observed data
}
